+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Fix ctrl\_user always set as root for mysql\_user resource
+
 ## 10.1.1 - *2021-03-11*
 
 - Fix db initialization status correctly on MySQL 8.0